SA-17-Clean up the code base && move background media to separate component
