+++ src/lib/components/modals/resultModal.svelte
