+++ src/lib/styles/global.vars.scss
@@ -12,8 +12,7 @@
 	// colors
 	--main-background-color: #2c2b2b;
 	--main-text-color: #eeeeee;
-	--main-second-text-color: #d1d0d0;
-	--main-second-background-color: #fff;
+
 	--main-alert-color: #de1324;
 	--main-input-color: #222;
